# app/resources/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOTR API",
        "description": "Description of LOTR API",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "url",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/equipments": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Retrieve a paginated list of equipment",
                "operationId": "f86b74aa4a259ea785d20e91f2141603",
                "responses": {
                    "200": {
                        "description": "A paginated list of equipment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to retrieve equipments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Store a newly created equipment in the database",
                "operationId": "c2839e2f416411b6d3355f47f1379f59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to create equipment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/equipments/{id}": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Retrieve a specific equipment by its ID",
                "operationId": "587bbc9b68d2f89e18e79a55b20443f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipment found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Update the specified equipment in the database",
                "operationId": "c6e9885e2346872856ec59f4f9d94378",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to update equipment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Soft delete the specified equipment in the database",
                "operationId": "d69a0ecf9f87a5332fb3083933b1fc4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Equipment deleted"
                    },
                    "400": {
                        "description": "Unable to delete equipment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/equipments/{id}/restore": {
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Restore a soft-deleted equipment",
                "operationId": "823696af11bb8b37d79ec431e0b2e6c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to restore",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipment restored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to restore equipment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/equipments/{id}/force": {
            "delete": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Permanently delete a soft-deleted equipment from the database",
                "operationId": "3bf8ccb19d463af14a538e586ad9dabe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the equipment to permanently delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Equipment permanently deleted"
                    },
                    "400": {
                        "description": "Unable to permanently delete equipment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EquipmentRequest": {
                "required": [
                    "name",
                    "type",
                    "made_by"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "Sword of Elendil"
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "Sword"
                    },
                    "made_by": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "Elven Smiths"
                    }
                },
                "type": "object"
            },
...
            "EquipmentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sword of Elendil"
                    },
                    "type": {
                        "type": "string",
                        "example": "Sword"
                    },
                    "made_by": {
                        "type": "string",
                        "example": "Elven Smiths"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}